FIX: Removed warnings.  Most of the warnings were due to a subclass not      overriding ALL versions of an overloaded function.  These were addressed      by filling in the rest of the API.  In some cases, a subclass was      overriding and overloading a function (changing input parameter to a      more specific type).  For these cases, we "hid" the original function      in the private section of the subclass.  There was one internal API      change.  SetInput/Output(i, vtkDataObject*) was changed to SetNthInput/Output(.)

+++ common/vtkPolyData.cxx
@@ -1344,14 +1344,14 @@ int vtkPolyData::ClipUpdateExtentWithWholeExtent()
   int valid = 1;
   int piece = this->GetUnstructuredUpdateExtent()->GetPiece();
   int numPieces = this->GetUnstructuredUpdateExtent()->GetNumberOfPieces();
-  int maxPieces;  
+  unsigned long maxPieces;  
   
   // Check to see if upstream filters can break up the data into the
   // requested number of pieces.
   maxPieces = this->GetUnstructuredInformation()->GetMaximumNumberOfPieces();
   if (numPieces > maxPieces)
     {
-    numPieces = maxPieces;
+    numPieces = (int) maxPieces;
     }
   
   if (numPieces <= 0 || piece < 0 || piece >= numPieces)
